Revert to using `uwsm app` for startup apps
For some reason, `uwsm-app` executed by hyprland at startup causes all
the requests to be piped at once to `uwsm-app-daemon-in`, which makes
the daemon process and return back only one request, discarding the
others. This is likely a bug in the uwsm-app daemon and/or in the
uwsm-app script itself.